spring-integration-security-ldaptive/src/main/java/org/bremersee/spring/security/ldaptive/authentication/UsernameToBindDnConverterProperty.java
/*
 * Copyright 2024 the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.bremersee.spring.security.ldaptive.authentication;

import java.util.function.Function;
import org.bremersee.spring.security.ldaptive.authentication.UsernameToBindDnConverter.ByDomainEmail;
import org.bremersee.spring.security.ldaptive.authentication.UsernameToBindDnConverter.ByUserRdnAttribute;

/**
 * The enum UsernameToBindDnConverterReference.
 *
 * @author Christian Bremer
 */
public enum UsernameToBindDnConverterProperty
    implements Function<LdaptiveAuthenticationProperties, UsernameToBindDnConverter> {

  /**
   * By user rdn attribute username to bind dn converter property.
   */
  BY_USER_RDN_ATTRIBUTE(ByUserRdnAttribute::new),

  /**
   * By domain email username to bind dn converter property.
   */
  BY_DOMAIN_EMAIL(ByDomainEmail::new);

  private final Function<LdaptiveAuthenticationProperties, UsernameToBindDnConverter> resolverFn;

  UsernameToBindDnConverterProperty(
      Function<LdaptiveAuthenticationProperties, UsernameToBindDnConverter> resolverFn) {
    this.resolverFn = resolverFn;
  }

  @Override
  public UsernameToBindDnConverter apply(LdaptiveAuthenticationProperties properties) {
    return resolverFn.apply(properties);
  }
}
